+++ packages/samples/headless-commerce-react/vitest.config.ts
@@ -1,0 +1,11 @@
+import {defineConfig} from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    exclude: ['**/e2e/**'],
+    environment: 'jsdom',
+  },
+  define: {
+    global: 'window',
+  },
+});