+++ packages/samples/headless-commerce-ssr-react-router/vite.config.ts
@@ -1,0 +1,11 @@
+import {reactRouter} from '@react-router/dev/vite';
+import {defineConfig} from 'vite';
+import tsconfigPaths from 'vite-tsconfig-paths';
+
+export default defineConfig({
+  plugins: [reactRouter(), tsconfigPaths()],
+  build: {
+    target: 'esnext',
+    outDir: 'dist',
+  },
+});