+++ scripts/deploy/execute-deployment-pipeline.mjs
@@ -22,35 +22,49 @@ function getVersionComposants(version) {
     parsedVersion?.major,
     parsedVersion?.minor,
     parsedVersion?.patch,
-    parsedVersion.prerelease[0] || undefined,
+    ...(parsedVersion?.prerelease[0] ? [parsedVersion.prerelease[0]] : []),
   ];
 }
 
 function getResolveVariableString(version, packageName) {
   const prNumber = process.env.PR_NUMBER;
   const versionComposantsOrdered = getVersionComposants(version);
-  if (process.env.PATCH_ONLY && prNumber) {
-    return `--resolve ${packageName}_PATCH_VERSION=${versionComposantsOrdered.concat(prNumber).join('.')}`;
-  } else {
-    return `
-    --resolve ${packageName}_MAJOR_VERSION=${versionComposantsOrdered.slice(0, 1)}
-    --resolve ${packageName}_MINOR_VERSION=${versionComposantsOrdered.slice(0, 2).join('.')}
-    --resolve ${packageName}_PATCH_VERSION=${versionComposantsOrdered.slice(0, 3).join('.')}
-  `;
-  }
+
+  // Use PR number as build if available and PATCH_ONLY is set
+  const {major, minor, patch} =
+    process.env.PATCH_ONLY && prNumber
+      ? {
+          major: '0',
+          minor: '0.0',
+          patch: versionComposantsOrdered
+            .slice(0, 3)
+            .concat(prNumber)
+            .join('.'),
+        }
+      : {
+          major: versionComposantsOrdered.slice(0, 1),
+          minor: versionComposantsOrdered.slice(0, 2).join('.'),
+          patch: versionComposantsOrdered.slice(0, 3).join('.'),
+        };
+
+  return `
+    --resolve ${packageName}_MAJOR_VERSION=${major} \
+    --resolve ${packageName}_MINOR_VERSION=${minor} \
+    --resolve ${packageName}_PATCH_VERSION=${patch} \
+  `.trim();
 }
 
 const root = getVersionComposants(rootJson.version);
-const IS_NIGHTLY = root.includes(undefined);
+const IS_NIGHTLY = root.length > 3;
 
 console.log(
   execSync(
     `
-  deployment-package package create --with-deploy
-    --version ${root.join('.')}
-    --resolve IS_NIGHTLY=${IS_NIGHTLY}
-    --resolve IS_NOT_NIGHTLY=${!IS_NIGHTLY}
-    ${getResolveVariableString(buenoJson.version, 'BUENO')}
+  deployment-package package create --with-deploy \
+    --version ${root.join('.')} \
+    --resolve IS_NIGHTLY=${IS_NIGHTLY} \
+    --resolve IS_NOT_NIGHTLY=${!IS_NIGHTLY} \
+    ${getResolveVariableString(buenoJson.version, 'BUENO')} \
     ${getResolveVariableString(headlessJson.version, 'HEADLESS')}
     ${getResolveVariableString(atomicJson.version, 'ATOMIC')}
     ${getResolveVariableString(atomicReactJson.version, 'ATOMIC_REACT')}